[Work] Remove unused components
