--- src/main/java/github/clone_code_detection/entity/highlight/HighlightMetadata.java
@@ -1,17 +1,0 @@
-package github.clone_code_detection.entity.highlight;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class HighlightMetadata {
-    String author;
-    String organization;
-    int semester;
-    String course;
-    String filename;
-    int year;
-    String project;
-    String assigner;
-}
