--- src/main/java/github/clone_code_detection/user/BaseUserDetails.java
@@ -1,6 +1,0 @@
-package github.clone_code_detection.user;
-
-public class BaseUserDetails {
-    private String id;
-    private String username;
-}
